+++ packages/@expo/cli/e2e/__tests__/export-embed-test.ts
@@ -108,7 +108,6 @@ it('runs `npx expo export:embed`', async () => {
     {
       env: {
         NODE_ENV: 'production',
-        EXPO_USE_STATIC: 'static',
         E2E_ROUTER_SRC: 'static-rendering',
         E2E_ROUTER_ASYNC: 'development',
       },
@@ -183,7 +182,6 @@ it('runs `npx expo export:embed --platform ios` with source maps', async () => {
     {
       env: {
         NODE_ENV: 'production',
-        EXPO_USE_STATIC: 'static',
         E2E_ROUTER_SRC: 'static-rendering',
         E2E_ROUTER_ASYNC: 'development',
       },
@@ -260,7 +258,6 @@ it('runs `npx expo export:embed --platform android` with source maps', async () 
     {
       env: {
         NODE_ENV: 'production',
-        EXPO_USE_STATIC: 'static',
         E2E_ROUTER_SRC: 'static-rendering',
         E2E_ROUTER_ASYNC: 'development',
       },
@@ -349,7 +346,6 @@ it('runs `npx expo export:embed --bytecode`', async () => {
     {
       env: {
         NODE_ENV: 'production',
-        EXPO_USE_STATIC: 'static',
         E2E_ROUTER_SRC: 'static-rendering',
         E2E_ROUTER_ASYNC: 'development',
       },